requirements.txt
asgiref==3.7.2
Django==4.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
mccabe==0.7.0
pep8-naming==0.13.2
pillow==10.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2023.3
sqlparse==0.4.4
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1